Split code into one file per type
